% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{addExtractionDirectives}
\alias{addExtractionDirectives}
\title{Extraction Directives}
\usage{
addExtractionDirectives(x, coverage = NULL, prefix = NULL, append = T, ...)
}
\arguments{
\item{x}{environment, a study}

\item{coverage}{numeric, a pair of values between 0 and 1 that specify the range
of coverage to be included in the category}

\item{prefix}{character, a prefix that will be added to each label in the form
prefix_label. It must not contain any underscore}

\item{append}{logic, initialize (FALSE) or append a new definition (TRUE)}

\item{...}{not implemented}
}
\description{
Training feature directives determine how training features (pixels) are extracted
from training data. \code{\link[=addExtractionDirectives]{addExtractionDirectives()}} either initialize or add a new
row to the table of directives.
}
\examples{
\dontrun{
addExtractionDirectives(MyStudy,c(0.75,100),'core', append = F)
addExtractionDirectives(MyStudy,c(0,0.75),'border', append = T)
}
}
