% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_methods.R
\name{addFilter}
\alias{addFilter}
\title{Filters list}
\usage{
addFilter(
  x,
  filter = NULL,
  parameters = NULL,
  channels = NULL,
  append = T,
  ...
)
}
\arguments{
\item{x}{environment, a study}

\item{filter}{character, type of filter can be 'vanvliet', 'deriche', 'blur_anisotropic', from the package imageR.}

\item{parameters}{list, list of named values to be passed to the specific function.}

\item{channels}{character vector, names of the layers to be processed.}

\item{append}{logical, FALSE delete the previously added filters, TRUE append the specified filters to the stack of filters already defined.}
}
\value{
an object of class IMC_filterFrame
}
\description{
\code{\link[=addFilter]{addFilter()}} either initialize or add a new filter to an object of class
IMC_FilterFrame. Filters are then processed via the command \code{\link[=deployFilters]{deployFilters()}}.
}
\examples{
\dontrun{
availableLayers<-ch_Rnames(MyStudy)
addFilter (x = MyStudy,
           filter = 'vanvliet',
           parameters = list(list(sigma=1,order=0,axis='x'), list(sigma=1,order=0,axis='y')),
           channels = availableLayers,
           append = F)
addFilter (x = MyStudy,
           filter = 'blur_anisotropic',
           parameters = list(list(amplitude=9),list(amplitude=3)),
           channels = availableLayers,
           append=T)
deployFilters (MyStudy, saveToDisk = T)
}

}
