% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alligatorMap.R
\name{alligatorMap}
\alias{alligatorMap}
\title{Some BullShit}
\usage{
alligatorMap(
  fn_srt,
  fn_radius = 10,
  fn_Nspikes = 4,
  fn_minArea = 10,
  fn_maxArea = 100,
  fn_minRoundness = 0.6,
  fn_maxRoundness = 0.8,
  fn_coverage = 0.3,
  fn_seedOutScore = 10,
  fn_cycleWindow = 1,
  fn_discoverTreshold = 0.01,
  fn_adaptative = T,
  fn_areaAdaptRate = 0,
  fn_roundnessAdaptRate = 0,
  fn_fusion = T,
  fn_targetArea = 50,
  fn_maxNetworkSize = 4,
  fn_inflateDeflate = 0.1,
  fn_favourForeing = T,
  fn_returnKinetic = F,
  fn_returnRasters = F
)
}
\description{
Some BullShit
}
