% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_methods.R
\name{an_dismissAnalysis}
\alias{an_dismissAnalysis}
\title{Dismiss an analysis}
\usage{
an_dismissAnalysis(x, analysis = NULL, ...)
}
\arguments{
\item{x}{environment, a study.}

\item{analysis}{character, name of the analysis to dismiss.
If NULL the \strong{currentAnalysis} will be dismissed.}
}
\value{
NULL.
}
\description{
Dismissing an analysis means removing the current or a previous
analysis from the specified study. Associated files are not
permanently removed but copied to a \strong{dismissed} subfolder in
the analysis folder of the specified study.
}
\examples{
\dontrun{
dismissAnalysis(MyStudy, 'MyUnwantedAnalysis')
}
}
