% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_methods.R
\name{an_showCurrentAnalysis}
\alias{an_showCurrentAnalysis}
\title{Show the current analysis}
\usage{
an_showCurrentAnalysis(x, ...)
}
\arguments{
\item{x}{environment, a study.}
}
\value{
print to the console a list of the current analysis objects ant their states.
}
\description{
Current analysis is an environment containing multiple objects.
\code{\link[=an_showCurrentAnalysis]{an_showCurrentAnalysis()}} lists the objects contained in the currentAnalysis
environment and assess whether each object is \strong{initialized} or \strong{empty}.
}
\examples{
\dontrun{
an_showCurrentAnalysis(MyStudy)
}
}
