% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_methods.R
\docType{methods}
\name{newAnalysis}
\alias{newAnalysis}
\alias{an_dismissAnalysis}
\alias{an_listAnalysis}
\alias{an_showCurrentAnalysis}
\title{Analysis}
\usage{
newAnalysis(x, analysisName = NULL, ...)

an_dismissAnalysis(x, analysis = NULL, ...)

an_listAnalysis(x, ...)

an_showCurrentAnalysis(x, ...)
}
\arguments{
\item{x}{environment, a study}

\item{analysisName}{character, a name for a new analysis}

\item{...}{not implemented}

\item{analysis}{character, name of the analysis to operate on}
}
\value{
an object of class environment within the specified study
}
\description{
Each study can have only one current analysis (myStudy$currentAnalysis) allocated in the environment
container currentAnalysis.
When creating a new analysis this take place of the currentent analysis (whenever it exists),
however an old analysis that has been archived can be retrieved as well.
}
\details{
an_dismissAnalysis: delete an analysis and the associated files

an_listAnalysis: list the analysis created in the specified study

an_showCurrentAnalysis: show a summary of the currentAnalysis
}
\examples{
\dontrun{
newAnalysis(x = MyStudy)
an_listAnalysis (x = MyStudy)
an_currentAnalysis (x = MyStudy)
an_dismissAnalysis (x = MyStudy) # dismiss current analysis
an_dismissAnalysis (x = Mystudy, analysis = 'example_analysis')
}
}
\seealso{

}
