% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadilloMap.R
\name{armadilloMap}
\alias{armadilloMap}
\title{Some BullShit}
\usage{
armadilloMap(
  fn_srt,
  fn_clpDir = c("8", "4"),
  fn_brake = 10,
  fn_lowerQuantile = 0.01,
  fn_upperQuantile = 0.5,
  fn_lowerAreaLimit = 1000,
  fn_meanArea = 100,
  fn_populationFactor = 5,
  fn_varianceArea = 1,
  fn_lineStringBuffer = 1,
  fn_criticalMass = 40,
  fn_PVoptimization = T,
  fn_PVBuffer = 1.3,
  fn_verbose = T,
  fn_radius = 10,
  fn_Nspikes = 4,
  fn_minArea = 10,
  fn_maxArea = 100,
  fn_minRoundness = 0.6,
  fn_maxRoundness = 0.8,
  fn_coverage = 0.3,
  fn_seedOutScore = 10,
  fn_cycleWindow = 1,
  fn_discoverTreshold = 0.01,
  fn_adaptative = T,
  fn_areaAdaptRate = 0,
  fn_roundnessAdaptRate = 0,
  fn_segmentAlg = c("sign_simple", "sign_extended", "quadratic"),
  fn_fusion = T,
  fn_targetArea = 50,
  fn_maxNetworkSize = 4,
  fn_inflateDeflate = 0.1,
  fn_favourForeing = T,
  fn_returnKinetic = F,
  fn_returnRasters = F
)
}
\description{
Some BullShit
}
