% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoTreshold.R
\name{autoTreshold}
\alias{autoTreshold}
\title{Some BullShit}
\usage{
autoTreshold(
  fn_df,
  fn_marker,
  fn_maxNOfEvents = 100,
  fn_tol = 1e-05,
  fn_expansion = 1,
  fn_seed = 1234
)
}
\description{
Some BullShit
}
