% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelTable_methods.R
\name{ch_markers}
\alias{ch_markers}
\title{Channel Table content}
\usage{
ch_markers(x, loaded = T, ...)
}
\arguments{
\item{x}{environment, a study.}

\item{loaded}{logical, which channel? loaded(T), unloaded(F), all(NA).}

\item{...}{not implemented.}
}
\value{
named character vector of names.
}
\description{
retrieve the parsed name of the markers.
}
\examples{
\dontrun{
default_study<-ch_markers(x = MyStudy, loaded = T)
}
}
