% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelTable_methods.R
\name{ch_Onames}
\alias{ch_Onames}
\title{Channel Table content}
\usage{
ch_Onames(x, loaded = T, ...)
}
\arguments{
\item{x}{environment, a study.}

\item{loaded}{logical, which channel? loaded(T), unloaded(F), all(NA).}

\item{...}{not implemented.}
}
\value{
named character vector of names.
}
\description{
retrive original channel names as they appear as column names in the original
raw data .txt file.
}
\examples{
\dontrun{
default_study<-ch_Onames(x = MyStudy, loaded = T)
}
}
