% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelTable_methods.R
\name{ch_RnameFromMarker}
\alias{ch_RnameFromMarker}
\title{Channel Table content}
\usage{
ch_RnameFromMarker(x, marker, ...)
}
\arguments{
\item{x}{environment, a study.}

\item{marker}{character, marker name.}

\item{...}{not implemented.}
}
\value{
named character.
}
\description{
retrieve the R compatible name of the channel corresponding to the specified
marker.
}
\examples{
\dontrun{
default_study<-ch_RnameFromMarker(x = MyStudy, marker = 'CD4')
}
}
