% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelTable_methods.R
\name{ch_Rnames}
\alias{ch_Rnames}
\title{Channel Table content}
\usage{
ch_Rnames(x, loaded = T, ...)
}
\arguments{
\item{x}{environment, a study.}

\item{loaded}{logical, which channel? loaded(T), unloaded(F), all(NA).}

\item{...}{not implemented.}
}
\value{
character vector of names.
}
\description{
retrive R compatible channel name. These names are used as layer names
for the raw data rasters.
}
\examples{
\dontrun{
default_study<-ch_Rnames(x = MyStudy, loaded = T)
}
}
