% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelTable_methods.R
\docType{methods}
\name{channelTable}
\alias{channelTable}
\alias{channelTable,environment-method}
\alias{ChannelTable,environment-method}
\alias{ch_Rnames}
\alias{ch_Rnames,environment-method}
\alias{ch_Rnames,environment}
\alias{ch_Onames}
\alias{ch_Onames,environment-method}
\alias{ch_Onames,environment}
\alias{ch_markers}
\alias{ch_markers,environment-method}
\alias{ch_markers,environment}
\alias{ch_RnameFromMarker}
\alias{ch_RnameFromMarker,environment-method}
\alias{ch_RnameFromMarker,environment}
\title{Channel Table content}
\usage{
channelTable(x, ...)

\S4method{channelTable}{environment}(x, ...)

ch_Rnames(x, loaded = T, ...)

\S4method{ch_Rnames}{environment}(x, loaded = T, ...)

ch_Onames(x, loaded = T, ...)

\S4method{ch_Onames}{environment}(x, loaded = T, ...)

ch_markers(x, loaded = T, ...)

\S4method{ch_markers}{environment}(x, loaded = T, ...)

ch_RnameFromMarker(x, marker, ...)

\S4method{ch_RnameFromMarker}{environment}(x, marker, ...)
}
\value{
an object of class IMC_ChannelTable or a character vector (named if different from R-columns)
}
\description{
Retrive or update content of the channel table. Some content such as R-names
are read only to prevent corruption of study cross references.
}
\details{
ChannelTable: retrive the entire study table, there is no assignment method

ch_Rnames: retrive R compatible names, loaded (T), unloaded (F), all (NA)

ch_Onames: retrive the original names in raw data TXT file, loaded (T), unloaded (F), all (NA)

ch_markers: retrive the original names in raw data TXT file, loaded (T), unloaded (F), all (NA)

ch_RnameFromMarker: retrive the R compatible name (usually use as layer name for raster objects) from the marker name
}
\examples{
\dontrun{
default_study<-channelTable(x = MyStudy, loaded = NA)
default_study<-channelTable(x = MyStudy, loaded = T)
ch_channel(x = myStudy)<-('MySample1','MySample2','MySample3')
}
}
\seealso{

}
