% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelTable_methods.R
\name{channelTable}
\alias{channelTable}
\title{Channel Table content}
\usage{
channelTable(x, ...)
}
\arguments{
\item{x}{environment, a study.}

\item{...}{not implemented.}
}
\value{
an object of class IMC_ChannelTable.
}
\description{
Retrive or update content of the channel table. Some content such as R-names
are read only to prevent corruption of study cross references.
}
\examples{
\dontrun{
default_study<-channelTable(x = MyStudy)
}
}
\seealso{

}
