% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_methods.R
\name{deployFilters}
\alias{deployFilters}
\title{Apply filters}
\usage{
deployFilters(x, saveToDisk = T, ...)
}
\arguments{
\item{x}{environment, a study.}

\item{saveToDisk}{logical, save new rasterStacks to disk? Always advisable.}
}
\value{
a new IMC_RsCollection that will be stored in the object \strong{derivedRasters}.
}
\description{
\code{\link[=deployFilters]{deployFilters()}} utilizes the filter definitions stored in the IMC_FilterFrame
object, created via the function \code{\link[=addFilter]{addFilter()}}.
This function is a wrapper for the internal \code{\link[=imageRFilter]{imageRFilter()}}.
}
\examples{
\dontrun{
availableLayers<-ch_Rnames(MyStudy)
addFilter (x = MyStudy,
           filter = 'vanvliet',
           parameters = list(list(sigma=1,order=0,axis='x'), list(sigma=1,order=0,axis='y')),
           channels = availableLayers,
           append = F)
addFilter (x = MyStudy,
           filter = 'blur_anisotropic',
           parameters = list(list(amplitude=9),list(amplitude=3)),
           channels = availableLayers,
           append=T)
deployFilters (MyStudy, saveToDisk = T)
}

}
