% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTiffImage.R
\name{.exportTiffImage}
\alias{.exportTiffImage}
\title{Export Tiff image}
\usage{
.exportTiffImage(
  fn_Raster = NULL,
  fn_Layers = NULL,
  fn_transpose = F,
  fn_flip = c("NA", "x", "y", "xy"),
  fn_scale = F,
  fn_path = NULL,
  fn_fileName = NULL,
  fn_bps = c("32", "16", "8"),
  fn_verbose = T
)
}
\arguments{
\item{fn_Raster}{IMC_RasterStack}
}
\value{

}
\description{
Export Tiff image
}
\examples{

\dontrun{
}

}
