% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTiffMask.R
\name{.importTiffMask}
\alias{.importTiffMask}
\title{Import tiff Mask}
\usage{
.importTiffMask(
  fn_fileList = NULL,
  fn_transpose = T,
  fn_layerName = "Tiff_imported"
)
}
\arguments{
\item{fn_fileList}{character vector, a list of tiff files to be imported}

\item{fn_transpose}{logical, transpose matrix? It should be if the raw data has been transposed as well}

\item{fn_layerName}{character, new name for the layer}
}
\value{
an object of class IMC_classification
}
\description{
Import a classification map from a different source \code{\link[=.importTiffMask]{.importTiffMask()}} is the
internal function for \code{\link[=importTiffMask]{importTiffMask()}} method.
}
\examples{

\dontrun{
importTiffMask(x = TEST,
fileFolder='c:/somewhere...',
orderAgain =c(1,3,2),
transposeImage = F,
layerName = 'Import',
saveToDisk = T ))
}

}
