% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTiffImage.R
\name{exportTiffImage}
\alias{exportTiffImage}
\title{Export Tiff image}
\usage{
exportTiffImage(
  x,
  layerName = NULL,
  fileFolder = NULL,
  transposeImage = F,
  flipImage = NULL,
  ...
)
}
\arguments{
\item{x}{environment, a study}

\item{fileFolder}{root folder containing at least a folder (which name is gone to be used as layer name), containing one tiff mask for each sample}

\item{...}{not implemented}
}
\value{
none
}
\description{
Export Tiff image
}
\examples{

\dontrun{
}

}
