% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{extractTrainingFeatures}
\alias{extractTrainingFeatures}
\title{Extraction of training features}
\usage{
extractTrainingFeatures(x, ...)
}
\arguments{
\item{x}{environment, a study}

\item{...}{not implemented}
}
\description{
Training feature (pixels) are extracted from training data, following
following the directives provided via the function \code{\link[=addExtractionDirectives]{addExtractionDirectives()}}.
}
\examples{
\dontrun{
addExtractionDirectives(MyStudy,c(0.75,100),'core', append = F)
extractTrainingFeatures(MyStudy)
}
}
