% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractXYPixels.R
\name{extractXYPixels}
\alias{extractXYPixels}
\title{Some BullShit}
\usage{
extractXYPixels(
  fn_polygons,
  fn_study,
  fn_whichSample,
  fn_whichPopulation,
  fn_howManyCells,
  fn_buffer = 10,
  fn_seed = 123
)
}
\description{
Some BullShit
}
