% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggp.biplot}
\alias{ggp.biplot}
\title{Plot IMC data and segmentation}
\usage{
ggp.biplot(
  fn_exprs = NULL,
  fn_type = c("scatter", "density"),
  fn_x = NULL,
  fn_y = NULL,
  fn_z = NULL,
  fn_xTrsh = NULL,
  fn_yTrsh = NULL,
  fn_zLimits = NULL,
  fn_limitsType = NULL,
  fn_xModulus = 0,
  fn_yModulus = 0,
  fn_colors = NULL,
  fn_title = NULL,
  fn_subtitle = NULL,
  fn_sep = "/",
  fn_bg = "white",
  fn_fg = "black",
  ...
)
}
\arguments{
\item{fn_exprs}{sf dataframe, eg. the table produced after \code{\link[=segment]{segment()}} and
\code{\link[=distillExpression]{distillExpression()}}.}

\item{fn_type}{character, plot type. Can be either scatter or density. For one
dimension, density is the only available.}

\item{fn_x}{character, column name to plot eleong x axis}

\item{fn_y}{same as fn_x}

\item{fn_z}{character, column name to be used as color in a scatter plot.}

\item{fn_xTrsh}{numeric, draw a vertical line as a threshold (like in flow cytometry).}

\item{fn_yTrsh}{same as fn_xTrsh}

\item{fn_zLimits}{numeric vector, boundaries for the dimension represented
as color gradient}

\item{fn_xModulus}{numeric, transformation to apply to x axis, \code{\link[scales:boxcox_trans]{scales::modulus_trans()}}.}

\item{fn_yModulus}{numeric, same as fn_xModulus}

\item{fn_colors}{character, Can be a single
color, a number of colors equal to the number of categories or a number of
colors to be used to construct a gradient in case of a continuous variable.}

\item{fn_title}{character, used as title}

\item{fn_subtitle}{character vector, a subtitle. A character vector will be
collapsed.}

\item{fn_sep}{character, character to use for collapse subtitle}

\item{fn_bg}{character, color to be used in background panel.}

\item{fn_fg}{character, color to be used in accessory graphics on the
foreground.}

\item{...}{not implemented}

\item{fn_limitisType}{character, determine how to use fn_zLimits. Can be one
among \strong{fix}, \strong{quantile}.}
}
\value{
ggplot
}
\description{
Simple scatter or density plot for one, two or three dimensions.
}
\examples{
\dontrun{

}

}
