% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggp.geometry}
\alias{ggp.geometry}
\title{Plot IMC data and segmentation}
\usage{
ggp.geometry(
  fn_gg = NULL,
  fn_geometry = NULL,
  fn_borderCol = NULL,
  fn_borderVar = NULL,
  fn_fillCol = NULL,
  fn_fillVar = NULL,
  fn_zoom = NULL,
  fn_title = NULL,
  fn_subtitle = NULL,
  fn_sep = "/",
  fn_bg = "white",
  ...
)
}
\arguments{
\item{fn_gg}{ggplot produced with \code{\link[=gg.raster]{gg.raster()}} to be overlayed.}

\item{fn_geometry}{sf dataframe, eg. the table produced after \code{\link[=segment]{segment()}} and
\code{\link[=distillExpression]{distillExpression()}}.}

\item{fn_borderCol}{character, colors to use for the border. Can be a single
color, a number of colors equal to the number of categories or a number of
colors to be used to construct a gradient in case of a continuous variable.}

\item{fn_borderVar}{character, name of the column to use to project colors
to the border}

\item{fn_fillCol}{character, the same as fn_borderCol applied to fill area.}

\item{fn_fillVar}{character, the same as fn_borderVar applied to fill area.}

\item{fn_zoom}{numeric vector, x and y limits of the area to plot.}

\item{fn_title}{character, title for the plot}

\item{fn_subtitle}{character vector, a subtitle. A character vector will be
collapsed.}

\item{fn_sep}{character, character to use for collapse subtitle}

\item{fn_bg}{character, color to be used in background panel.}

\item{...}{not implemented}
}
\value{
ggplot
}
\description{
\code{\link[=ggp.geometry]{ggp.geometry()}} can be used to plot the result of segmentation as polygons,
projecting on both border and area any of the continuos of discrete variables
contained in the exprs matrix.
In conjunction with \code{\link[=gg.raster]{gg.raster()}} overlays polygon outlines over a raster image.
}
\examples{
\dontrun{

sampleToUse <- 1
uidTouse<-st_uids(MyStudy)[sampleToUse]

ggp<-ggp.raster(fn_raster = MyStudy$raster[[uidTouse]]$X191Ir.DNA1.Ir191Di.,
  fn_limits = c(0,0.99),
  fn_limitsType = 'quantile',
  fn_zoom = c(100,150,100,150),
  fn_title = uidTouse,
  fn_subtitle = c(st_samples(MyStudy)[sampleToUse],st_rois(MyStudy)[sampleToUse]))

ggp<-ggp.geometry(fn_gg = ggp,
  fn_geometry = myStudy$currentAnalysis$exprs[
    MySudy$currentAnalysis$exprs$uid == uidToUse, ],
  fn_borderVar = 'label',
  fn_zoom = c(100,150,100,150),
  fn_title = st_uids(MyStudy)[1],
  fn_subtitle = c(st_samples(MyStudy)[1],st_rois(MyStudy)[1]),
  )

print(ggp)
}
}
