% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggp.raster}
\alias{ggp.raster}
\title{Plot IMC data and segmentation}
\usage{
ggp.raster(
  fn_raster = NULL,
  fn_colors = NULL,
  fn_limits = NULL,
  fn_limitsType = NULL,
  fn_zoom = NULL,
  fn_title = NULL,
  fn_subtitle = NULL,
  fn_sep = "/",
  fn_bg = "white",
  ...
)
}
\arguments{
\item{fn_raster}{rasterLayer.}

\item{fn_colors}{character vector, colors to be used to construct the gradiant
used for plotting th raster. If NULL, black to white will be used.
Specifying only one color a gradient from black will be constructed.}

\item{fn_limits}{numeric or character vector, limits of the color gradient,
everything outside will be squished. Permitted values depends on the fn_limitsType
specified.}

\item{fn_limitsType}{character vector, one among \strong{fix}, \strong{quantile} or
\strong{category}.}

\item{fn_zoom}{numeric vector, x and y limits of the area to plot.}

\item{fn_title}{character, title for the plot}

\item{fn_subtitle}{character vector, a subtitle. A character vector will be
collapsed.}

\item{fn_sep}{character, character to use for collapse subtitle}

\item{fn_bg}{character, color to be used in background panel.}

\item{...}{not implemented}

\item{fn_gg}{ggplot to be overlayed.}
}
\value{
ggplot
}
\description{
\code{\link[=ggp.raster]{ggp.raster()}} can be used to plot a rater image for a single marker.
The image can be overlayed with the product of segmentation using the
function \code{\link[=ggp.geometry]{ggp.geometry()}}.
}
\examples{
\dontrun{
ggp<-ggp.raster(fn_raster = MyStudy$raster[[1]]$X191Ir.DNA1.Ir191Di.,
  fn_limits = c(0,0.99),
  fn_limitsType = 'quantile',
  fn_zoom = c(100,150,100,150),
  fn_title = st_uids(MyStudy)[1],
  fn_subtitle = c(st_samples(MyStudy)[1],st_rois(MyStudy)[1]))

print(ggp)
}
}
