% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAllClasses.R
\docType{class}
\name{IMC_ExtractionDirectives-class}
\alias{IMC_ExtractionDirectives-class}
\alias{IMC_ExtractionDirectives}
\title{A Class container directives}
\description{
A Class container directives
}
