% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{IMC_RsCollection-class}
\alias{IMC_RsCollection-class}
\alias{IMC_RsCollection}
\title{A Class container for IMC_RasterStaks
Namely a collection of images from different histological preparations.}
\description{
A Class container for IMC_RasterStaks
Namely a collection of images from different histological preparations.
}
