% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{IMC_SegmentationDirectives-class}
\alias{IMC_SegmentationDirectives-class}
\alias{IMC_SegmentationDirectives}
\title{A Class container directives}
\description{
A Class container directives
}
