% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{IMC_SegmentationList-class}
\alias{IMC_SegmentationList-class}
\alias{IMC_SegmentationList}
\title{A Class container for the results of segmentation}
\description{
A Class container for the results of segmentation
}
