% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAllClasses.R
\docType{class}
\name{IMC_TrainingFeatures-class}
\alias{IMC_TrainingFeatures-class}
\alias{IMC_TrainingFeatures}
\title{A class container for treaning features}
\description{
A class container for treaning features
}
