% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTiffMask.R
\name{importTiffMask}
\alias{importTiffMask}
\title{Import tiff Mask}
\usage{
importTiffMask(
  x,
  fileFolder = NULL,
  orderAgain = NULL,
  transposeImage = T,
  layerName = "Tiff_imported",
  saveToDisk = T,
  ...
)
}
\arguments{
\item{x}{environment, a study}

\item{fileFolder}{root folder containing at least a folder (which name is gone to be used as layer name), containing one tiff mask for each sample}

\item{orderAgain}{numeric vector, which mask match which raw data file. Consider that list.files produce a list of files in sttrict alphabetical order}

\item{transposeImage}{logical, parameter passed to fn_transpose}

\item{layerName}{character, parameter passed to fn_layerName}

\item{saveToDisk}{logical, save rater to disk? absolutely for large data sets}

\item{...}{not implemented}
}
\value{
an object of class IMC_classification
}
\description{
Import a classification map from a different source.
Uses the internal \code{\link[=.importTiffMask]{.importTiffMask()}}.
}
\examples{

\dontrun{
importTiffMask(x = TEST,
fileFolder='c:/somewhere...',
orderAgain =c(1,3,2),
transposeImage = F,
layerName = 'Import',
saveToDisk = T ))
}

}
