% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initStudy.R
\name{initStudy}
\alias{initStudy}
\title{Some BullShit}
\usage{
initStudy(
  fn_studyName = "IMCstudy",
  fn_rootFolder = NULL,
  fn_rawDataFolder = NULL,
  fn_whichColumns = "named",
  fn_overWrite = F,
  ...
)
}
\description{
Some BullShit
}
