% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTxtToStack.R
\name{loadTxtToStack}
\alias{loadTxtToStack}
\title{Internal function to convert a txt file to a RasterStack usually executed during study initialisation.}
\usage{
loadTxtToStack(
  fn_path,
  fn_file,
  fn_rasterDBPath = NULL,
  fn_rasterStackPath = NULL,
  fn_cols = NULL,
  fn_newNames = NULL,
  fn_norm = F,
  fn_details = list(study = "NO_study", sample = "NO_sample", ROI = "NO_ROI", replicate
    = "NO_replicate", bioGroup = "No_bioGroup"),
  fn_channel = NULL,
  fn_trsh = 0.965,
  fn_zeroOff = T
)
}
\arguments{
\item{fn_path}{Root path for a bunch of .txt image files}
}
\description{
Internal function to convert a txt file to a RasterStack usually executed during study initialisation.
}
