% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localCorrection.R
\name{localCorrection,environment-method}
\alias{localCorrection,environment-method}
\title{Clean up Classification}
\usage{
\S4method{localCorrection}{environment}(
  x,
  labelLayer = "label",
  suffix = "_clean",
  matrixExtent = 3,
  paddingLabel = "undetermined",
  ...
)
}
\arguments{
\item{x}{environment, the study whom classification is to modify}

\item{labelLayer}{character, the name of the layer to modify. Must be a discrete
value layer.}

\item{suffix}{character, suffix to add to labelLayer to form the new classification
layer name.}

\item{matrixExtent}{numeric, extent of the square matrix to use in the focal
function, must be an odd number.}

\item{paddingLabel}{character, a class to be used for padding. Must be a class
already present in the classification to modify.}
}
\value{
an IMC_classification object that will overwrite the current analysis
classification
}
\description{
Method used to 'clean' the classification layer. Essentially a wrapper around
\code{\link[raster:focal]{raster::focal()}} that use a function that substitute the classification of
a pixel with the most abundant one in the square matrix defined by the parameter
matrixExtent.
}
