% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_methods.R
\name{newAnalysis}
\alias{newAnalysis}
\title{Create a new Analysis}
\usage{
newAnalysis(x, analysisName = NULL, ...)
}
\arguments{
\item{x}{environment, a study.}

\item{analysisName}{character, name of the new analysis. This name needs to be
a valid R name and a valid name for a directory. The same name cannot be used
twice within the same study.}
}
\value{
a new analysis environment in MyStudy$currentAnalysis.
}
\description{
An IMC analysis is composed of few objects contained in the environment
\strong{currentAnalysis} and a corresponding sub-folder of the study folder.
\code{\link[=newAnalysis]{newAnalysis()}} create an empty analysis structure both in the specified
study environment and on disk, essentially overwriting the analysis object
but keeping the file structure, so a previous analysis can always be loaded
as current via the \code{\link[=retrieve]{retrieve()}} method.
}
\examples{
\dontrun{
newAnalysis(MyStudy, 'MyFirstAnalysis')
}
}
