% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandaMap.R
\name{pandaMap}
\alias{pandaMap}
\title{Some BullShit}
\usage{
pandaMap(
  fn_srt = NULL,
  fn_uid = NULL,
  fn_primaryIndex = "PI",
  fn_clpDir = c("8", "4"),
  fn_brake = 10,
  fn_lowerQuantile = 0.01,
  fn_upperQuantile = 0.5,
  fn_lowerAreaLimit = 5,
  fn_movingWindow_dim = NULL,
  fn_movingWindow_overlap = 25:50,
  fn_cores = parallel::detectCores(logical = T) - 1,
  fn_verbose = F
)
}
\description{
Some BullShit
}
