% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotRaster.R
\name{plot.raster}
\alias{plot.raster}
\title{Some BullShit}
\usage{
plot.raster(
  fn_raster = NULL,
  fn_layer = NULL,
  fn_colorsContinuous = NULL,
  fn_scaleColourLimits = NULL,
  fn_limitsType = NULL,
  fn_colorsDiscrete = NULL,
  fn_labels = NULL,
  fn_segmentation = NULL,
  fn_segmentLabel = NULL,
  fn_segmentColor = NULL,
  fn_zoom = NULL
)
}
\description{
Some BullShit
}
