% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsPoly.R
\name{plotsPoly}
\alias{plotsPoly}
\title{Some BullShit}
\usage{
plotsPoly(
  fn_rst,
  fn_xmin,
  fn_xmax,
  fn_ymin,
  fn_ymax,
  fn_plcn,
  fn_plc,
  fn_plcRange,
  fn_xaxs,
  fn_yaxs,
  fn_xaxt,
  fn_yaxt,
  fn_colNA,
  fn_Bx,
  fn_By,
  fn_bgc,
  fn_geom,
  fn_geomB,
  fn_title
)
}
\description{
Some BullShit
}
