% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\docType{methods}
\name{retrieve}
\alias{retrieve}
\title{Retrieve}
\usage{
retrieve(fn_file = NULL, ...)
}
\arguments{
\item{fn_file}{character, file path of the study or part of it to be loaded from disk}
}
\description{
retrieve is used to load from disk an entire study, an Analysis or a part of it
}
\details{
retrieve use file extension to determine how to load the specified file
}
\examples{
\dontrun{
MyStudy<-retrieve(x = "c:/Documents/MyStudy.xml")
MyStudy$currentAnalysis<-retrieve(x= 'c:/Documents/MySudy/Analysis/Analysis1.xml')
}
}
\seealso{

}
