% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slothMap.R
\name{slothMap}
\alias{slothMap}
\title{Some BullShit}
\usage{
slothMap(
  fn_srt,
  fn_radius = 10,
  fn_Nspikes = 4,
  fn_minArea = 10,
  fn_maxArea = 100,
  fn_minRoundness = 0.6,
  fn_maxRoundness = 0.8,
  fn_coverage = 0.3,
  fn_seedOutScore = 10,
  fn_cycleWindow = 1000,
  fn_discoverTreshold = 0.01,
  fn_adaptative = T,
  fn_drastic = 0,
  fn_lowPenalty = 2,
  fn_highPenalty = 1,
  fn_roundnessPenalty = 1,
  fn_areaAdaptRate = 0,
  fn_roundnessAdaptRate = 0,
  fn_fusion = T,
  fn_areaMean = 50,
  fn_seed = 1234
)
}
\description{
Some BullShit
}
