% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFeaturesManipulation.R
\name{tf_labelList}
\alias{tf_labelList}
\title{Training features}
\usage{
tf_labelList(x, ...)
}
\arguments{
\item{x}{environment, IMC_TrainingFeatures, a study or the object storing training features}
}
\value{
numeric vector
}
\description{
Lists labels present in the training features table
}
