% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFeaturesManipulation.R
\name{tf_perimeterQuantile}
\alias{tf_perimeterQuantile}
\title{Training features}
\usage{
tf_perimeterQuantile(x, quantiles, ...)
}
\arguments{
\item{x}{environment, IMC_TrainingFeatures, a study or the object storing training features}

\item{quantiles}{numeric, vector of quantiles to be calculated}
}
\value{
numeric matrix
}
\description{
Calculates specified quantiles of perimeter for each label.
}
