% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFeaturesManipulation.R
\name{tf_roundnessStatistics}
\alias{tf_roundnessStatistics}
\title{Training features}
\usage{
tf_roundnessStatistics(x, ...)
}
\arguments{
\item{x}{environment, IMC_TrainingFeatures, a study or the object storing training features}

\item{quantiles}{numeric, vector of quantiles to be calculated}
}
\value{
numeric matrix
}
\description{
Calculates roundness summary statistics for each label.
}
