% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoMap.R
\name{topoMap}
\alias{topoMap}
\title{Some BullShit}
\usage{
topoMap(
  fn_rstStack = NULL,
  fn_layerLabel = NULL,
  fn_label = NULL,
  fn_area = NULL,
  fn_prefix = "topoMap_",
  fn_raster = NULL,
  fn_derivedRaster = NULL,
  fn_features = NULL,
  fn_maxClumps = 10,
  fn_clumpDirection = 8,
  fn_mtry = length(fn_features)/3,
  fn_ntree = 100,
  fn_trace = 10
)
}
\description{
Some BullShit
}
