% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMetadata.R
\name{updateMetadata}
\alias{updateMetadata}
\title{updateMetadata}
\arguments{
\item{x}{environment, representing a study}

\item{saveToDisk}{logic, save updated rasterStaks and raster collection}
}
\value{
the method act in place on the correct object within the environment specified
}
\description{
Update metadata in raster collections such as the raw data or classification rasters.
}
\examples{
\dontrun{
updateMetadata( x = MyStudy, saveToDisk = T)
}

}
