% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderMap.R
\name{spiderMap}
\alias{spiderMap}
\title{Some BullShit}
\usage{
spiderMap(
  fn_srt,
  fn_interpret,
  fn_radius = 10,
  fn_densityMultiplier = 1,
  fn_Nspikes = 4,
  fn_minArea = 10,
  fn_maxArea = 100,
  fn_minRoundness = 0.6,
  fn_maxRoundness = 0.8,
  fn_coverage = 0.3,
  fn_seedOutScore = 10,
  fn_cutSeedList = 0.1,
  fn_cycleWindow = 1000,
  fn_discoverTreshold = 0.01,
  fn_adaptative = T,
  fn_drastic = 0,
  fn_direction = "random",
  fn_seed = 1234
)
}
\description{
Some BullShit
}
